requirements.txt
openai
python-dotenv
pillow
pygame
speechrecognition
pyaudio